Random exercises

+++ C/Cube/main.c
@@ -5,9 +5,9 @@
 int main()
 {
 	short edgeLength;
-	int volume;
+	double volume;
 	printf("Entrez la longueur de l'arete du cube\n");
 	scanf("%d", &edgeLength);
 	volume = pow(edgeLength, 3);
-	printf("Le volume du cube est de %d\n", volume);
+	printf("Le volume du cube est de %e\n", volume);
 }
